Changed download method

+++ update.ps1
@@ -6,6 +6,9 @@ if ($PSVersionTable.PSVersion.Major -lt 3) {
     return
 }
 
+$wc = New-Object System.Net.WebClient
+$wc.Encoding = [System.Text.Encoding]::UTF8
+
 Write-Host "Updating application..."
 
 If (Test-Path "$current\assets") {
@@ -18,7 +21,7 @@ If (Test-Path "$current\update.ps1") {
     Remove-Item -Recurse -Force "$current\update.ps1" | Out-Null
 }
 
-Invoke-WebRequest -Uri "https://github.com/braathen/qlik-hardware-configurator/archive/master.zip" -UseBasicParsing -OutFile "./master.zip"
+$wc.DownloadFile("https://bit.ly/qhc-update", "$current\master.zip")
 
 Add-Type -assembly "system.io.compression.filesystem"
 
@@ -26,7 +29,9 @@ Add-Type -assembly "system.io.compression.filesystem"
 
 Move-Item "$current/qlik-hardware-configurator-master/*" -Force
 
-Remove-Item -Recurse -Force "$current\.gitignore"
+If (Test-Path "$current\.gitignore") {
+    Remove-Item -Recurse -Force "$current\.gitignore"
+}
 
 Remove-Item -Recurse -Force "$current\qlik-hardware-configurator-master"
